// src/interfaces/Imovel/comissao.tsx
type PlanilhasHistoricoType = {
  active: 0 | 1
  arquivo: string
  core_request_id: string | null
  id: number | string
  identifica_documento: "comissão"
  identifica_documento_id: number | string  
  nome_original: string
  parcela_id: number | string
  processo_id: number | string
  tipo_documento_id: null
  created_at: string 
}

type ParcelasUserComissionType = {
  "porcentagem_real": string | null,
  "valor_real": string | null,
  "usuario_id": string | number,
  "comissao_id": string | number,
  "empresa_id": string | null,
  "parcela_id": string | number,
  "total_comissao": string | null,
  "name": string | null,
  "cpf": string | null,
  "desconto": string | null,
  "tipo_pessoa": "PF" | "PJ",
  "nome_empresarial": string | null,
  "cnpj": string | null,
  "tipo_pagamento": "pix" | "banco",
  "agencia": string | null,
  "numero_conta": string | null,
  "pix": string | null,
  "banco_id": string | number,
  "nome_banco": string | null,
  "creci": string | null,
  "tipo_chave_pix_id": string | null,
  "chave_pix": string | null
  tipo_laudo_opcionista: string | null
  tipo_laudo_opcionista_label: string | null
}

type ParcelasComissionType = {
  "id": string | number,
  "processo_id": string | number,
  "comissao_id": string | number,
  "periodo_pagamento": string | null,
  "dia_pagamento": string | null,
  "numero_parcela": string | null,
  "valor_parcela": string | null,
  "ultima_data_envio": string | null,
  "data_criacao": string | null,
  "data_ultima_atualizacao": string | null,
  "corretores_vendedores": ParcelasUserComissionType[],
  "corretores_opcionistas": ParcelasUserComissionType[],
  "gerentes": ParcelasUserComissionType[],
  "gerentes_gerais": ParcelasUserComissionType[],
  "diretores_gerais": ParcelasUserComissionType[],
  "repasse_franquias": ParcelasUserComissionType[],
  "royalties": ParcelasUserComissionType,
  "comunicacao": ParcelasUserComissionType,
  "empresas": ParcelasUserComissionType[],
  "planilha_id": string   
  planilhas: PlanilhasHistoricoType[]
}